Fix postgres database restore to not require some magic database
Problem: No database name was given when running `psql` to terminate
other clients and drop and create the target database. It seems that
Postgres then defaults to choose a default database name based on the
user name, which results in errors if no such database exists.

Solution: Pass explicitly the database name `postgres` when running
commands to operate on databases.